initScraper: remove dataAuth,  postDataAuth & start. Register on portal
Removed unused methods, this could be used on portal's older versions.
register on portal behavior is: if the user session is not registered on the portal, but it exists, then display an html form with authentication data, that form is expected to be submitted as is. After that, when get the main page, the portal will return an HTML without any form but only the links to access recibidos / emitidos.